Move files encryption password retrieval to a separate method. Check if account is public.
The code handling password checks for encrypted files has been moved from a mixed context in the Init function to its own function, GetFilesPassword, improving the modular structure of the code. Furthermore, the condition checking whether passwords are needed for account and files has been adjusted to cover this new separate handling. The change marks an overall cleaner and more manageable structure of password checking in the ConnectionManager class.